Add Messages with Headers and Body (#3)
* Define the Created Header

* Define a Message with Headers and a Body

* Use Message within MessageChannel

* Rename CallBack as Channel for imporved semantics

User now make calls referring to a Channel rather than a Callback,
which is not an term used in the EIP book.

The type alias Channel has been renamed as ChannelHandle to make way.

* [readme] Updated

* [readme] add badges

* [Message] make default constructor private

* [Message] add javadoc

* [MessageChannel] Add javadoc

* [MessageChannel] Drop channels that don't need any environment

If the Sender or the Receiver don't need an Environment then they
if they do anything, then they can't be pure. So, to clarify, remove
the types that support that impure usage.